pyproject.toml
[project]
name = "mcap_manager"
version = "0.1.0"
description = "A CLI tool for querying and managing MCAP files"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "mcap>=0.3.0",
    "click>=8.1.0",
    "python-dateutil>=2.8.2",
    "tqdm>=4.65.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
mcap_manager = "mcap_manager.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 